+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/components/parsed/GeographicArea.vue
@@ -9,76 +9,78 @@
       klass="CollectingEvent"
       @selected="selectGeographicArea"
     />
-      <div v-if="areasByCoors.length">
-        <h4>By coordinates</h4>
-        <p>
-          <label>
-            <input
-              type="radio"
-              :checked="areasByCoors[0].id == collectingEvent.geographic_area_id"
-              @click="selectGeographicArea(areasByCoors[0])"
-            >
-            <span v-html="areasByCoors[0].label_html" />
-          </label>
-        </p>
-        <a
-          v-if="areasByCoors.length > 1"
-          class="cursor-pointer"
-          @click="showModal = true"
-        >
-          Show other options
-        </a>
-      </div>
-
-      <modal-component
-        v-if="showModal"
-        @close="showModal = false"
+    <div v-if="areasByCoors.length">
+      <h4>By coordinates</h4>
+      <p>
+        <label>
+          <input
+            type="radio"
+            :checked="areasByCoors[0].id == collectingEvent.geographic_area_id"
+            @click="selectGeographicArea(areasByCoors[0])"
+          >
+          <span v-html="areasByCoors[0].label_html" />
+        </label>
+      </p>
+      <a
+        v-if="areasByCoors.length > 1"
+        class="cursor-pointer"
+        @click="showModal = true"
       >
-        <template #header>
-          <h3>
-            Select geographic area
-          </h3>
-        </template>
-        <template #body>
-          <ul class="no_bullets">
-            <li
-              class="separate-bottom"
-              v-for="item in areasByCoors"
-              :key="item.id"
-            >
-              <label>
-                <input
-                  type="radio"
-                  :checked="item.id == collectingEvent.geographic_area_id"
-                  @click="selectGeographicArea(item); showModal = false"
-                >
-                <span v-html="item.label_html" />
-              </label>
-            </li>
-          </ul>
-        </template>
-      </modal-component>
-    <template v-if="geographicArea">
-      <div class="middle separate-top">
-        <span data-icon="ok" />
-        <span class="separate-right"> {{ geographicArea.name }}</span>
-        <span
-          class="circle-button button-default btn-undo"
-          @click="selectGeographicArea()"
-        />
-      </div>
-    </template>
+        Show other options
+      </a>
+    </div>
+
+    <modal-component
+      v-if="showModal"
+      @close="showModal = false"
+    >
+      <template #header>
+        <h3>
+          Select geographic area
+        </h3>
+      </template>
+      <template #body>
+        <ul class="no_bullets">
+          <li
+            class="separate-bottom"
+            v-for="item in areasByCoors"
+            :key="item.id"
+          >
+            <label>
+              <input
+                type="radio"
+                :checked="item.id == collectingEvent.geographic_area_id"
+                @click="selectGeographicArea(item); showModal = false"
+              >
+              <span v-html="item.label_html" />
+            </label>
+          </li>
+        </ul>
+      </template>
+    </modal-component>
+    <SmartSelectorItem
+      v-if="geographicArea"
+      :item="geographicArea"
+      label="name"
+      @unset="selectGeographicArea()"
+    />
+    <MetaPrioritizeGeographicArea
+      :disabled="!geographicArea"
+      v-model="collectingEvent.meta_prioritize_geographic_area"
+    />
   </fieldset>
 </template>
 
 <script>
-
+import { MutationNames } from '../../store/mutations/mutations'
 import { GetterNames } from '../../store/getters/getters'
 import { ActionNames } from '../../store/actions/actions'
 import { GeographicArea } from 'routes/endpoints'
 import SmartSelector from 'components/ui/SmartSelector.vue'
 import convertDMS from 'helpers/parseDMS.js'
 import ModalComponent from 'components/ui/Modal'
+import MetaPrioritizeGeographicArea from '../Meta/MetaPrioritizeGeographicArea.vue'
+import SmartSelectorItem from 'components/ui/SmartSelectorItem.vue'
 
 import extendCE from '../mixins/extendCE'
 
@@ -87,7 +89,9 @@ export default {
 
   components: {
     SmartSelector,
-    ModalComponent
+    ModalComponent,
+    MetaPrioritizeGeographicArea,
+    SmartSelectorItem
   },
 
   computed: {
@@ -105,6 +109,15 @@ export default {
 
     geographicArea () {
       return this.$store.getters[GetterNames.GetGeographicArea]
+    },
+
+    collectingEvent: {
+      get () {
+        return this.$store.getters[GetterNames.GetCollectingEvent]
+      },
+      set (value) {
+        this.$store.commit(MutationNames.SetCollectingEvent, value)
+      }
     }
   },
 
@@ -148,8 +161,8 @@ export default {
       if (this.collectingEvent.geographic_area_id) return
       if (convertDMS(this.verbatimLatitude) && convertDMS(this.verbatimLongitude)) {
         clearTimeout(this.ajaxCall)
-        this.ajaxCall = setTimeout(() => { 
-          this.getByCoords(convertDMS(this.verbatimLatitude), convertDMS(this.verbatimLongitude)) 
+        this.ajaxCall = setTimeout(() => {
+          this.getByCoords(convertDMS(this.verbatimLatitude), convertDMS(this.verbatimLongitude))
         }, this.delay)
       }
     }