+++ app/javascript/vue/tasks/collection_objects/browse/components/Panel/PanelDerived.vue
@@ -1,0 +1,16 @@
+<template>
+  <TableGrid
+    :columns="2"
+    :column-width="{ default: 'minmax(100px, auto)' }"
+    gap="1em"
+  >
+    <PanelDwc />
+    <PanelTimeline />
+  </TableGrid>
+</template>
+
+<script setup>
+import PanelTimeline from './PanelTimeline.vue'
+import PanelDwc from './PanelDwc.vue'
+import TableGrid from 'components/layout/Table/TableGrid.vue'
+</script>