+++ app/javascript/vue/tasks/collection_objects/browse/store/actions/loadTimeline.js
@@ -1,0 +1,6 @@
+import { CollectionObject } from 'routes/endpoints'
+
+export default ({ state }, coId) =>
+  CollectionObject.timeline(coId).then(({ body }) => {
+    state.timeline = body
+  })