+++ app/models/identifier/local.rb
@@ -22,8 +22,8 @@
 #
 class Identifier::Local < Identifier
 
-  # This must exist, rather than namespace: true, because we don't have database side not null in the model.  We also don't accept nested
-  # namespaces in this belongs to.
+  # This must exist, rather than namespace: true, because we don't have database side not null in the model. We also don't accept nested
+  # namespaces in this belongs_to.
   validates :namespace_id, presence: true
 
   validates_uniqueness_of :identifier, scope: [:namespace_id, :project_id, :type], message: lambda { |error, attributes| "#{attributes[:value]} already taken"}
@@ -72,10 +72,6 @@ class Identifier::Local < Identifier
     delimiter = '' if delimiter == 'NONE'
 
     [namespace&.verbatim_short_name, namespace&.short_name, ''].compact.first + delimiter
-  end
-
-  def set_cached
-    update_column(:cached, build_cached)
   end
 
 end