+++ app/models/source/bibtex.rb
@@ -407,25 +407,38 @@ class Source::Bibtex < Source
     p
   end
 
-  # @return [BibTeX::Entry]
+  # @return [Source::Bibtex.new]
+  #   Adds errors if parse error exists. Note these
+  # errors are lost if save/valid? is called again on the object.
   def self.new_from_bibtex_text(text = nil)
-    a = BibTeX::Bibliography.parse(text, filter: :latex).first
-    new_from_bibtex(a)
+    source = Source::Bibtex.new
+    begin
+      a = BibTeX::Bibliography.parse(text, filter: :latex).first
+      if a.class.name == 'BibTeX::Error'
+        source.errors.add(:base, 'Unable to parse BibTeX entry. Possible error at end of: ' + a.content)
+        return source
+      else
+        return new_from_bibtex(a)
+      end
+    rescue BibTeX::ParseError => e
+      source.errors.add(:base, 'Unable to parse BibTeX entry: ' + e.to_s)
+      return source
+    end
   end
 
   # Instantiates a Source::Bibtex instance from a BibTeX::Entry
   # Note:
-  #    * note conversion is handled in note setter.
-  #    * identifiers are handled in associated setter.
-  #    * !! Unrecognized attributes are added as import attributes.
+  #   * note conversion is handled in note setter.
+  #   * identifiers are handled in associated setter.
+  #   * !! Unrecognized attributes are added as import attributes.
   #
   # Usage:
-  #    a = BibTeX::Entry.new(bibtex_type: 'book', title: 'Foos of Bar America', author: 'Smith, James', year: 1921)
-  #    b = Source::Bibtex.new_from_bibtex(a)
+  #   a = BibTeX::Entry.new(bibtex_type: 'book', title: 'Foos of Bar America', author: 'Smith, James', year: 1921)
+  #   b = Source::Bibtex.new_from_bibtex(a)
   #
   # @param [BibTex::Entry] bibtex_entry the BibTex::Entry to convert
-  # @return [Source::BibTex.new] a new instance
-  # TODO annote to project specific note?
+  # @return [Source::Bibtex.new] a new instance
+  # TODO: Annote to project specific note?
   # TODO: Serial with alternate_value on name .count = 1 assign .first
   def self.new_from_bibtex(bibtex_entry = nil)
     return false if !bibtex_entry.kind_of?(::BibTeX::Entry)
@@ -434,6 +447,8 @@ class Source::Bibtex < Source
     import_attributes = []
 
     bibtex_entry.fields.each do |key, value|
+      next if key == :serial # Raises if it hits the belongs_to
+
       if key == :keywords
         s.verbatim_keywords = value
         next
@@ -629,18 +644,11 @@ class Source::Bibtex < Source
       s = Serial.where(name: journal).first
       i = Identifier::Global::Issn.where(identifier: value, identifier_object_type: 'Serial').first
 
-      # Found an Existing Serial identically named with
-      # an Identical ISSN
-      if s && i
+      # Found an Existing Serial identically named with an assigned Identical ISSN
+      if !s.nil? && (s == i&.identifier_object)
         write_attribute(:serial_id, s.id)
-      elsif i && s.nil? # Found an identifier, but not on a identically named Serial, assign it anyways
+      elsif i && s.nil? # Found a Serial with an Identifier, but not identically named, assign it anyway
         write_attribute(:serial_id, i.identifier_object_id)
-      elsif i.nil? && s.nil? && !journal.blank? && !value.blank?
-        # Found neither, but provided ISSN and Serial, this will get caught in new_from_bibtex to create a new serial
-      else
-        # Some other strange combination has occurred, add some erorrs (not tested)
-        errors.add(:journal, 'Conflict between ISSN and article type')
-        errors.add(:issn, 'Conflict between ISSN and article type')
       end
     end
   end
@@ -653,7 +661,7 @@ class Source::Bibtex < Source
   # turn bibtex URL field into a Ruby URI object
   # @return [URI]
   def url_as_uri
-    URI(self.url) unless self.url.blank?
+    URI(self.url) if self.url.present?
   end
 
   # @param [String] type_value
@@ -704,7 +712,7 @@ class Source::Bibtex < Source
   # @return [Integer]
   #  The effective year of publication as per nomenclatural rules
   def nomenclature_year
-    cached_nomenclature_date.year
+    cached_nomenclature_date&.year
   end
 
   # @return [Date || Time] <sigh>